IMP dir struct and copyright
