+++ base_field_validator/views/ir_model_view.xml
